requirements.txt
Flask==2.1.0
picamera2
py-mjpeg